Allow configuring the javahome
The javahome to use when linking can now be selected with `--javahome`.
This path will be used to discover the classes from the JRE that will be
available to the program at runtime.

Previously, this path would be given via the `--classpath` option, but
this doesn't work well for Java 9+, because there are no JARs ready to
be read.