Optionally write dot graph of dependencies
This commit introduces a new option `--dot-output <path>` which defines
the path where to write a dot graph representing the dependencies
between methods, classes or packages of the application.

Whether the graph should represent dependencies between methods, classes
or packages can be configured with `--dot-granularity <granularity>`.

The packages included in the produced dot graph can be filtered using
`--dot-package-filter <package>`. When set, symbols belonging to
`<package>` or any of its subpackages will be included in the dot graph.