Discover mains from JAR using Manifest
This commit adds a new option, `--entrypoint-from-manifest`, which can
be used to specify that the given JAR file's manifest should be read,
and the main class that it defines should be added as entrypoint.

This is useful when the JAR contains other main methods that are
runnable, but which we're not interested in testing.